make find shell command path customizable
On powershell, there already exists a find shell command, which is incompatible with this makefile.
If it comes first in path, the makefile fails. One can edit the path variable, but may not wish to do so for other reasons.
One needs to use a posix compliant find command instead, as can be found from Git, which is provided as an example.

Also, removing trailing spaces at various positions.